add "disableable" module
A very simple module that checks a flag on the user's model.  If the
flag is true then the account is disabled, the user's session will be
ended and the user will no longer be able to sign in.  The user can be
re-enabled by setting the flag to false or nil.